refactor(LottoManager): 랜덤 로또를 만드는 메서드의 위치를 변경
LottoManager의 생성전에 랜덤으로 생성한 로또를 사용자에게 출력해야 한다.

BREAKING CHANGE:
    LottoManager의 makeRandomLottoList메서드가 Lotto도메인으로 이동한다.
