+++ src/test/java/lotto/LottoManagerTest.java
@@ -13,39 +13,6 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueN
 public class LottoManagerTest {
 
     @Test
-    void 시도에_맞춰서_랜덤_로또를_알맞은_개수를_뽑아야_한다() {
-        LottoManager lottoManager = new LottoManager(
-                new Attempt(4000),
-                new Lotto(List.of(1,2,3,4,5,6)),
-                new BonusLotto(7));
-
-        long expected = 4;
-
-        Assertions.assertEquals(expected,lottoManager.getRandomLottoList().size());
-    }
-
-    @Test
-    void 랜덤_로또를_알맞게_뽑아야_한다() {
-        List<Lotto> expected = getRandomList().stream()
-                .map(Lotto::new)
-                .toList();
-
-        assertRandomUniqueNumbersInRangeTest(
-                () -> {
-                    LottoManager lottoManager = new LottoManager(
-                            new Attempt(4000),
-                            new Lotto(List.of(1,2,3,4,5,6)),
-                            new BonusLotto(7));
-                    Assertions.assertEquals(expected,lottoManager.getRandomLottoList());
-                },
-                getRandomList().get(0),
-                getRandomList().get(1),
-                getRandomList().get(2),
-                getRandomList().get(3)
-        );
-    }
-
-    @Test
     void 보너스_적용이_제대로_작동해야_한다() {
         int expected = 1;
 
@@ -54,7 +21,8 @@ public class LottoManagerTest {
                     LottoManager lottoManager = new LottoManager(
                             new Attempt(1000),
                             new Lotto(List.of(1,2,3,4,5,7)),
-                            new BonusLotto(1));
+                            new BonusLotto(1),
+                            Lotto.makeRandomLottoList(1));
                     Assertions.assertEquals(
                             expected,
                             lottoManager.doLotto().get(LottoPrize.FIVE_MATCH_BONUS)
@@ -72,7 +40,8 @@ public class LottoManagerTest {
                     LottoManager lottoManager = new LottoManager(
                             new Attempt(2000),
                             new Lotto(List.of(1,2,3,4,5,6)),
-                            new BonusLotto(1));
+                            new BonusLotto(1),
+                            Lotto.makeRandomLottoList(2));
                     Assertions.assertEquals(
                             expected,
                             lottoManager.doLotto().get(LottoPrize.SIX_MATCH)
@@ -91,7 +60,8 @@ public class LottoManagerTest {
                     LottoManager lottoManager = new LottoManager(
                             new Attempt(1000),
                             new Lotto(List.of(1,2,3,4,5,6)),
-                            new BonusLotto(1));
+                            new BonusLotto(1),
+                            Lotto.makeRandomLottoList(1));
                     Assertions.assertEquals(
                             expected,
                             lottoManager.doLotto()
@@ -117,7 +87,8 @@ public class LottoManagerTest {
                     LottoManager lottoManager = new LottoManager(
                             new Attempt(usedMoney),
                             new Lotto(lotto),
-                            new BonusLotto(bonusNumber));
+                            new BonusLotto(bonusNumber),
+                            Lotto.makeRandomLottoList(2));
                     Map<LottoPrize, Integer> prizeMap = lottoManager.doLotto();
                     Assertions.assertEquals(
                             expected,
@@ -127,14 +98,6 @@ public class LottoManagerTest {
                 fiveMatchBonus,
                 sixMatch
         );
-    }
-
-    private List<List<Integer>> getRandomList() {
-        return List.of(
-                List.of(8, 21, 23, 41, 42, 43),
-                List.of(3, 5, 11, 16, 32, 38),
-                List.of(7, 11, 16, 35, 36, 44),
-                List.of(1, 8, 11, 31, 41, 42));
     }
 
     private Map<LottoPrize, Integer> initializePrizeMap() {
